+++ kernel/peak_errno.c
@@ -20,6 +20,7 @@ const char *peak_errno_str(int err) {
     case PEAK_EBUSY: return "busy";
     case PEAK_EDHCP: return "DHCP failed";
     case PEAK_EAGAIN: return "try again";
+    case PEAK_ELOOP: return "too many symlink levels";
     default: return "unknown error";
     }
 }